[PORT FROM MAIN][BT] Firmware file different on ffrd and on vv
BZ: 103648

In bt_vendor_xxx.conf, the path we now set for the Bluetooth firmware files is
not the final path any more but a 'root' path containing a vv/ and a pr/
folder.
At run-time, depending on whether we are on a vv board or on a form factor,
firmwares will be seeked in the correct directory.

Notes:
 - for now, we rely on the ro.spid.wifi.nvram property to know if we are on
a vv or a ffrd. This will soon be abandoned by Wi-Fi. This property should be
renamed then.
 - For TI chips, the firmware file location is not set here

Category: device enablement
Domain: CWS.BT-Core
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: I0fb33f2839448fc5423212c67116919cce6189c9
Orig-Change-Id: Ic15985896b0d97e53da3a4ed960f1fec0694b39f
Signed-off-by: Xavier Boniface <xavierx.boniface@intel.com>
Reviewed-on: http://android.intel.com:8080/105036
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/123807
Reviewed-by: Poignet, Cedric <cedric.poignet@intel.com>
Tested-by: cactus <cactus@intel.com>