[BT] Add support for multiple bt_vendor.conf at runtime
BZ: 117478

To select the bt_vendor.conf at runtime, it is required to build the
system image with the bt_vendor.conf for every platforms we may want to
support.
Therefore the files are not more named bt_vendor.conf on the target image
but bt_baylake.conf, bt_redhookbay.conf, etc...

The right bt_"vendor".conf is informed by a property "ro.bt.conf_file"
which can be filled at boot time.

Change-Id: Ib66ac9a516a8ec44d6b261a12014cecf5d4912f0
Signed-off-by: Patrick Combes <patrick.combes@intel.com>
Reviewed-on: http://android.intel.com:8080/126529
Reviewed-by: Fert, Laurent <laurent.fert@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>