[BT] Fix Audio parameter value type
BZ: 170548

*_set_param() functions get a p_value pointer on a character.
It was missused as an integer value.
That's why the ASCII value was printed instead of the integer one.

This patch cleans the use the parameter value.
It fixes a compilation warning about the HW_SCO_I2S case missed in the switch too.

Change-Id: I8d328534d3248a08ae3c9c68f13c980ba5947ccd
Category: device enablement
Domain: CWS.BT-BCM
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: Cedric Poignet <cedric.poignet@intel.com>