Retry to bind on the Service
It's possible that the System service is not constructed and so
that the Binder is not registered before the Bluedroid Initialization.
It results in a binding failure and only a BT OFF ON can recover from
the issue.

This patch starts a thread in case the first bind failed and sleep for
1 second then retry after 1 more second for each tentative.

Change-Id: I2e130f129d24349275d314a6b3a435122de33f2f
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-16958
Signed-off-by: Cedric Poignet <cedric.poignet@intel.com>
Reviewed-on: https://android.intel.com:443/328896